DEV-22 fixed tests and coverage
